refactor(game): controller 패키지명을 game으로 변경
역할은 controller이지만 프로젝트 방향성에 부합하는 game으로 패키지명 변경
GameResult 얼겨형 클래스명을 Result로 변경
